[receiver/fluentforward] fix logs pipeline and convert multiple spans
The support for parsing spans was recently added in the
fluentforwardreceiver. The change moved the logic of reading the tag
from the DecodeMsg interface implementations to first reading the tag
and determining whether the event is a log or a span, then parsing it
accordingly. Because of this, we need to add these fields
(array length and tag) to the DecodeMsg interface.

This commit also adds the ability to convert multiple spans to the
internal format, which was missing.

Signed-off-by: Aditya <aditya10699@gmail.com>